+++ src/app/pwa-management/reminders/reminders.service.ts
@@ -1,72 +1,111 @@
-import { updateTask } from './../../../../backend/handlers';
 import { inject, Injectable } from '@angular/core';
 import { SwPush } from '@angular/service-worker';
-import { IdbService } from '../../api-management/idb-service/idb.service';
-import { from } from 'rxjs';
+import { from, isEmpty } from 'rxjs';
+import { TaskService } from '../../api-management/interceptors/task.service';
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
-
 export class RemindersService {
-  db = inject(IdbService)
-  swPush = inject(SwPush)
-
-  requestPermission() {
+  db = inject(TaskService);
+  swPush = inject(SwPush);
+  timeOut!: ReturnType<typeof setTimeout>;
+  delay!: number;
+  /** Get request to show notifications */
+  requestPermission(): void {
     from(Notification.requestPermission()).subscribe({
       next: (permission: NotificationPermission) => {
-        if (permission == "granted") {
-          console.log("Notification permission granted.");
-          this.getNextReminder()
+        if (permission == 'granted') {
+          console.log('Notification permission granted.');
+          this.getNextReminder();
         }
-        if (permission == "denied") {
-          console.warn("Notification permission denied.");
+        if (permission == 'denied') {
+          console.warn('Notification permission denied.');
         }
-      }
-    })
+      },
+    });
   }
 
-  sendNotification(data: Reminder) {
-    if (Notification.permission === "granted") {
+  /** Send a system level notification */
+  sendNotification(data: Reminder): void {
+    if (Notification.permission === 'granted') {
       new Notification(data.title, {
         body: data.subtext,
-        icon: "icons/icon-96x96.png",
+        icon: 'icons/icon-96x96.png',
         requireInteraction: true,
         silent: false,
       });
     }
   }
 
+  /** Finds the next available reminder and sets a timeout for that */
   getNextReminder(): void {
+    if (this.timeOut) {
+      clearTimeout(this.timeOut);
+    }
     this.db.getTasks().subscribe({
       next: (tasks) => {
         const now = Date.now();
-        const nextObj = []
+        const nextObj = [];
         for (let task of tasks) {
           if (task.dueDate) {
-            const time = new Date(task.dueDate).getTime()
-            time > now ? nextObj.push(task) : undefined
+            const time = new Date(task.dueDate).getTime();
+            time > now ? nextObj.push(task) : undefined;
           }
         }
-        const futureReminders = tasks.filter((task) => task.dueDate ? new Date(task.dueDate as string).getTime() > now : false)
-        const nextReminder = futureReminders.sort((a, b) => a.dueDate ? new Date(a.dueDate as string).getTime() - new Date(b.dueDate as string).getTime() : NaN)[0] || null;
-        const delay = new Date(nextReminder.dueDate as string).getTime() - Date.now()
-        if (delay > 0) {
-          console.log(`Next notification in ${delay / 1000} seconds`);
-          setTimeout(() => {
-            this.playNotificationSound()
-            this.sendNotification({
-              title: nextReminder.title,
-              subtext: nextReminder.description,
-              date: nextReminder.dueDate as string,
-            });
-          }, delay);
+        const futureReminders = tasks.filter((task) =>
+          task.dueDate
+            ? new Date(task.dueDate as string).getTime() > now
+            : false
+        );
+        if (futureReminders.length == 0) {
+          return;
         }
-      }
-    })
+        const nextReminder =
+          futureReminders.sort((a, b) =>
+            a.dueDate
+              ? new Date(a.dueDate as string).getTime() -
+                new Date(b.dueDate as string).getTime()
+              : NaN
+          )[0] || null;
+        this.delay =
+          new Date(nextReminder.dueDate as string).getTime() - Date.now();
+        if (this.delay > 0) {
+          this.applyReminder(nextReminder);
+        }
+      },
+    });
   }
 
-  playNotificationSound() {
+  /** Apply reminder based on task */
+  setNextReminder(task: Task): void {
+    const newDelay = new Date(task.dueDate as string).getTime() - Date.now();
+    if (this.delay) {
+      if (newDelay < this.delay) {
+        clearTimeout(this.timeOut);
+      }
+    }
+    this.delay = newDelay;
+    this.applyReminder(task);
+  }
+
+  /** Logic for setting notification timer */
+  private applyReminder(task: Task) {
+    console.log(`Next notification in ${this.delay / 1000} seconds`);
+    this.timeOut = setTimeout(() => {
+      this.playNotificationSound();
+      this.sendNotification({
+        title: task.title,
+        subtext: task.description,
+        date: task.dueDate as string,
+      });
+    }, this.delay);
+  }
+
+  /** Uses audio file to simulate notification sound */
+  playNotificationSound(): void {
     const audio = new Audio('/notification.mp3');
-    audio.play().catch(error => console.error("Audio playback error:", error));
+    audio
+      .play()
+      .catch((error) => console.error('Audio playback error:', error));
   }
 }