+++ src/app/types/reminder.d.ts
@@ -1,1 +1,2 @@
+/** Notification type */
 type Reminder = { title: string; subtext: string;  date: string}