+++ public/changelog.md
@@ -1,3 +1,10 @@
+## 0.0.4
+
+1. Set up to always sync and overwrite remote DB when online
+2. Setup to always add copy into indexedDB even when online
+3. Notification setup when adding tasks
+4. Clearing previous timers if the new task was due earlier
+
 ## 0.0.3
 
 1. added system notification and sound management for db durations